+++ src/js/sdk.tsx
@@ -1,9 +1,7 @@
 /* eslint-disable complexity */
 import type { Scope } from '@sentry/core';
-import { getIntegrationsToSetup, hasTracingEnabled, Hub, initAndBind, makeMain, setExtra } from '@sentry/core';
-import { HttpClient } from '@sentry/integrations';
+import { getIntegrationsToSetup, Hub, initAndBind, makeMain, setExtra } from '@sentry/core';
 import {
-  defaultIntegrations as reactDefaultIntegrations,
   defaultStackParser,
   getCurrentHub,
   makeFetchTransport,
@@ -11,24 +9,9 @@ import {
 import type { Integration, UserFeedback } from '@sentry/types';
 import { logger, stackParserFromStackParserOptions } from '@sentry/utils';
 import * as React from 'react';
-import { Platform } from 'react-native';
 
 import { ReactNativeClient } from './client';
-import {
-  DebugSymbolicator,
-  DeviceContext,
-  EventOrigin,
-  HermesProfiling,
-  ModulesLoader,
-  ReactNativeErrorHandlers,
-  ReactNativeInfo,
-  Release,
-  SdkInfo,
-} from './integrations';
-import { NativeLinkedErrors } from './integrations/nativelinkederrors';
-import { createReactNativeRewriteFrames } from './integrations/rewriteframes';
-import { Screenshot } from './integrations/screenshot';
-import { ViewHierarchy } from './integrations/viewhierarchy';
+import { getDefaultIntegrations } from './integrations/default';
 import type { ReactNativeClientOptions, ReactNativeOptions, ReactNativeWrapperOptions } from './options';
 import { ReactNativeScope } from './scope';
 import { TouchEventBoundary } from './touchevents';
@@ -39,11 +22,6 @@ import { getDefaultEnvironment } from './utils/environment';
 import { safeFactory, safeTracesSampler } from './utils/safe';
 import { NATIVE } from './wrapper';
 
-const IGNORED_DEFAULT_INTEGRATIONS = [
-  'GlobalHandlers', // We will use the react-native internal handlers
-  'TryCatch', // We don't need this
-  'LinkedErrors', // We replace this with `NativeLinkedError`
-];
 const DEFAULT_OPTIONS: ReactNativeOptions = {
   enableNativeCrashHandling: true,
   enableNativeNagger: true,
@@ -102,50 +80,9 @@ export function init(passedOptions: ReactNativeOptions): void {
     options.environment = getDefaultEnvironment();
   }
 
-  const defaultIntegrations: Integration[] = passedOptions.defaultIntegrations || [];
-  if (passedOptions.defaultIntegrations === undefined) {
-    defaultIntegrations.push(new ModulesLoader());
-    if (Platform.OS !== 'web') {
-      defaultIntegrations.push(new ReactNativeErrorHandlers({
-        patchGlobalPromise: options.patchGlobalPromise,
-      }));
-    }
-    defaultIntegrations.push(new Release());
-    defaultIntegrations.push(...[
-      ...reactDefaultIntegrations.filter(
-        (i) => !IGNORED_DEFAULT_INTEGRATIONS.includes(i.name)
-      ),
-    ]);
-
-    defaultIntegrations.push(new NativeLinkedErrors());
-    defaultIntegrations.push(new EventOrigin());
-    defaultIntegrations.push(new SdkInfo());
-    defaultIntegrations.push(new ReactNativeInfo());
-
-    if (__DEV__) {
-      defaultIntegrations.push(new DebugSymbolicator());
-    }
-
-    defaultIntegrations.push(createReactNativeRewriteFrames());
-    if (options.enableNative) {
-      defaultIntegrations.push(new DeviceContext());
-    }
-    if (options._experiments && typeof options._experiments.profilesSampleRate === 'number') {
-      defaultIntegrations.push(new HermesProfiling());
-    }
-    if (hasTracingEnabled(options) && options.enableAutoPerformanceTracing) {
-      defaultIntegrations.push(new ReactNativeTracing());
-    }
-    if (options.attachScreenshot) {
-      defaultIntegrations.push(new Screenshot());
-    }
-    if (options.attachViewHierarchy) {
-      defaultIntegrations.push(new ViewHierarchy());
-    }
-    if (options.enableCaptureFailedRequests) {
-      defaultIntegrations.push(new HttpClient());
-    }
-  }
+  const defaultIntegrations: false | Integration[] = passedOptions.defaultIntegrations === undefined
+    ? getDefaultIntegrations(options)
+    : passedOptions.defaultIntegrations;
 
   options.integrations = getIntegrationsToSetup({
     integrations: safeFactory(passedOptions.integrations, { loggerMessage: 'The integrations threw an error' }),